Moved 'authorizeTFA' function to Login module
