+++ src/modules/TFA.js
@@ -1,5 +1,5 @@
 import speakeasy from 'speakeasy';
-import { AuthenticationError, UserInputError } from 'apollo-server-core';
+import { AuthenticationError, UserInputError } from 'apollo-server-express';
 import QRCode from 'qrcode';
 import { User } from '../models';
 import jsonwebtoken from 'jsonwebtoken';
@@ -54,33 +54,6 @@ export const getBackupCodes = async (payload) => {
   if (!user) throw new AuthenticationError('Must be logged in');
 
   return user.backupCodes;
-};
-
-export const authorizeTFA = async (payload, backupCode, token) => {
-  if (!payload) throw new AuthenticationError('Must have gone through the first factor auth');
-  if (!payload.requiresTFA) throw new UserInputError('User does not need to authorize TFA');
-
-  if (backupCode) {
-    // TODO maybe it's possible to do everything in one query?
-    var { backupCodes, name, username } = await User.findById(
-      payload.id,
-      'backupCodes name username'
-    );
-    const codeIndex = backupCodes.findIndex((code) => code.code === token && !code.used);
-    if (codeIndex === -1) throw new AuthenticationError('Invalid backup code');
-    backupCodes[codeIndex].used = true;
-    await User.updateOne({ _id: payload.id }, { backupCodes });
-  } else {
-    var { tfa, name, username } = await User.findById(payload.id, 'tfa name username');
-
-    const verified = speakeasy.totp.verify({ secret: tfa, encoding: 'base32', token, window: 2 });
-    if (!verified) throw new AuthenticationError('Invalid token');
-  }
-  return {
-    token: jsonwebtoken.sign({ id: payload.id, name, username }, process.env.JWT_SECRET, {
-      expiresIn: payload.remember ? '1y' : '1d',
-    }),
-  };
 };
 
 const generateBackupCodes = () => {
