+++ src/modules/index.js
@@ -1,3 +1,3 @@
-export { login } from './Login';
+export { login, authorizeTFA } from './Login';
 export { signup } from './Signup';
-export { generateSecret, enableTFA, disableTFA, getBackupCodes, authorizeTFA } from './TFA';
+export { generateSecret, enableTFA, disableTFA, getBackupCodes } from './TFA';
