+++ src/resolvers/Login.js
@@ -1,8 +1,9 @@
-import { login } from '../modules';
+import { login, authorizeTFA } from '../modules';
 
 const loginResolvers = {
   Mutation: {
     login: (_, args) => login(args.email, args.password, args.remember),
+    authorizeTFA: (_, { token, backupCode }, { user }) => authorizeTFA(user, backupCode, token),
   },
 };
