+++ src/resolvers/TFA.js
@@ -1,4 +1,4 @@
-import { generateSecret, enableTFA, disableTFA, getBackupCodes, authorizeTFA } from '../modules';
+import { disableTFA, enableTFA, generateSecret, getBackupCodes } from '../modules';
 
 const tfaResolvers = {
   Query: {
@@ -6,7 +6,6 @@ const tfaResolvers = {
     getBackupCodes: (_, _args, { user }) => getBackupCodes(user),
   },
   Mutation: {
-    authorizeTFA: (_, { token, backupCode }, { user }) => authorizeTFA(user, backupCode, token),
     enableTFA: (_, { secret, token }, { user }) => enableTFA(user, secret, token),
     disableTFA: (_, { token }, { user }) => disableTFA(user, token),
   },
